requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0

# Machine Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
detectron2==0.6
transformers==4.35.2

# 3D Processing
open3d==0.18.0
trimesh==4.0.5
pyvista==0.42.3

# Video Processing
ffmpeg-python==0.2.0
imageio==2.31.5
imageio-ffmpeg==0.4.9

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0
requests==2.31.0
aiofiles==23.2.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
